+++ content/src/main/content/jcr_root/apps/acs-commons/components/utilities/redirectmappage/clientlibs/app.js
@@ -2,7 +2,7 @@
  * #%L
  * ACS AEM Commons Bundle
  * %%
- * Copyright (C) 2017 Adobe
+ * Copyright (C) 2015 Adobe
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,33 +18,44 @@
  * #L%
  */
 /*global angular: false */
-angular.module('acs-commons-automatic-package-replication-app', ['acsCoral', 'ACS.Commons.notifications'])
+angular.module('acs-commons-redirectmappage-app', ['acsCoral', 'ACS.Commons.notifications'])
     .controller('MainCtrl', ['$scope', '$http', '$timeout', 'NotificationsService',
     function ($scope, $http, $timeout, NotificationsService) {
 
-        $scope.save = function () {
+        $scope.updateRedirectMap = function () {
         	
 			NotificationsService.running(true);
-            
 
-            var $form = $('#fn-acsCommons-APR-form');
+            var $form = $('#fn-acsCommons-update-redirect');
+
+            $.ajax({
+                url: $form.attr('action'),
+                data: new FormData($form[0]),
+                cache: false,
+                contentType: 'multipart/form-data',
+                processData: false,
+                type: 'POST',
+                success: function(data){
+                	location.reload(true);
+                }
+            });
+            return false;
+        };
+        
+        $scope.postValues = function (id) {
+        	
+			NotificationsService.running(true);
+
+            var $form = $('#'+id);
 
             $.post($form.attr('action'), $form.serialize(), function() {
-               location.reload(true);
+            	location.reload(true);
             });
+            return false;
         };
 
         $scope.init = function () {
-        	$('#trigger-select').change(function(){
-            	$('#event-container').hide();
-            	$('#cron-container').hide();
-            	if($('#trigger-select').val() == 'event'){
-            		$('#event-container').show();
-            	}
-            	if($('#trigger-select').val() == 'cron'){
-            		$('#cron-container').show();
-            	}
-            });
+        	
         };
     }]);